Bài tập anh Tuấn

+++ MinhDuc/h1n-bai1.cpp
@@ -3,33 +3,20 @@
 
 int main ()
 {
-int a, b, c;
+long long a, b, c;
 	
       printf("nhap gia tri a b c :  ");
-      scanf("%d%d%d",&a,&b,&c);
+      scanf("%lld%lld%lld",&a,&b,&c);
 	int max;
 	  max = a;
-	if(max < b)
-	{
-	  max = b;
-	}
-	if(max < c)
-	{
-	  max = c;
-	}
-	  printf("\nMax la: %d", max);
+	if(max < b) {max=b;}
+	if(max < c) {max=c;}
+	  printf("\nMax la: %lld", max);
 	int min;
-	
 	  min = a;
-	if(min > b)
-	{
-	  min = b;
-	}
-	if(min > c)
-	{
-	  min = c;
-	}
-	printf("\nMin la: %d", min);
+	if(min > b) {min=b;}
+	if(min > c) {min=c;}
+	printf("\nMin la: %lld", min);
 
 return 0;
 }
